fass_updated/faas/packages/speech-to-text/speech-to-text/requirements.txt

google-cloud-speech==2.16.2

